+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - ALPHA: Added Contacts extension - shipped disabled
 - ALPHA: Added Calendar extension - shipped disabled
 - Added force re-sync of contacts
+- Added separate smtp credentials option - [#264](https://github.com/hkdb/aerion/issues/264)
+- Added no outgoing server option - [#132](https://github.com/hkdb/aerion/issues/132) [(#134)](https://github.com/hkdb/aerion/pull/134)
 - Added fallback for Mailfence and other non-quote-compliant providers - [#209](https://github.com/hkdb/aerion/issues/209)
 - Fixed CardDav remove provider code path to not leave orphaned contacts in db
 - Fixed Sent/Draft folder message listing - [#227](https://github.com/hkdb/aerion/issues/227)